Extract IgniteStartStopProcessorInitTest
